# Conflicts:
#	src/main/java/com/dace/dmgr/combat/Combat.java
#	src/main/java/com/dace/dmgr/event/listener/OnPlayerQuit.java
#	src/main/java/com/dace/dmgr/lobby/User.java
#	src/main/java/com/kiwi/dmgr/GameUser.java
#	src/main/java/com/kiwi/dmgr/match/MatchMaking.java
#	src/main/java/com/kiwi/dmgr/match/RankRating.java

+++ src/main/java/com/dace/dmgr/DMGR.java
@@ -1,10 +1,10 @@
 package com.dace.dmgr;
 
 import com.dace.dmgr.combat.event.CombatEventManager;
-import com.dace.dmgr.config.GeneralConfig;
 import com.dace.dmgr.event.MainEventManager;
-import com.dace.dmgr.lobby.Lobby;
 import com.dace.dmgr.lobby.User;
+import com.dace.dmgr.system.EntityInfoRegistry;
+import com.dace.dmgr.system.SystemPrefix;
 import com.dace.dmgr.system.command.LobbyCommand;
 import com.dace.dmgr.system.command.MenuCommand;
 import com.dace.dmgr.system.command.PlayerOptionCommand;
@@ -17,8 +17,6 @@ import com.kiwi.dmgr.game.map.WorldManager;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
-
-import static com.dace.dmgr.system.HashMapList.userMap;
 
 /**
  * 플러그인 메인 클래스.
@@ -38,8 +36,7 @@ public class DMGR extends JavaPlugin {
      */
     @Override
     public void onEnable() {
-        getLogger().info(PREFIX.LOG + "플러그인 활성화 완료");
-        GeneralConfig.init();
+        getLogger().info(SystemPrefix.LOG + "플러그인 활성화 완료");
         MainEventManager.init();
         CombatEventManager.init();
         registerCommands();
@@ -50,13 +47,11 @@ public class DMGR extends JavaPlugin {
 
         Bukkit.getOnlinePlayers().forEach((Player player) -> {
             User user = new User(player);
-            userMap.put(player, user);
-            Lobby.lobbyTick(player);
-            getServer().broadcastMessage(PREFIX.CHAT + "플레이어 할당 : §e§n" + player.getName());
+            user.init();
+            getServer().broadcastMessage(SystemPrefix.CHAT + "플레이어 등록 : §e§n" + player.getName());
         });
-        Bukkit.getOnlinePlayers().forEach((Player player) -> {
-            player.sendMessage(PREFIX.CHAT + "시스템 재부팅 완료");
-        });
+        Bukkit.getOnlinePlayers().forEach((Player player) ->
+                player.sendMessage(SystemPrefix.CHAT + "시스템 재부팅 완료"));
     }
 
     /**
@@ -64,12 +59,13 @@ public class DMGR extends JavaPlugin {
      */
     @Override
     public void onDisable() {
-        getLogger().info(PREFIX.LOG + "플러그인 비활성화 완료");
+        getLogger().info(SystemPrefix.LOG + "플러그인 비활성화 완료");
 
         Bukkit.getOnlinePlayers().forEach((Player player) -> {
-            User user = new User(player);
-            user.getLobbySidebar().delete();
-            player.sendMessage(PREFIX.CHAT + "시스템 재부팅 중...");
+            User user = EntityInfoRegistry.getUser(player);
+            if (user != null)
+                user.getLobbySidebar().delete();
+            player.sendMessage(SystemPrefix.CHAT + "시스템 재부팅 중...");
         });
     }
 
@@ -90,17 +86,5 @@ public class DMGR extends JavaPlugin {
         getCommand("소환").setExecutor(new DummyCommand());
         getCommand("게임테스트").setExecutor(new GameTestCommand());
         getCommand("스텟").setExecutor(new StatCommand());
-    }
-
-    /**
-     * 시스템 메시지의 접두사.
-     */
-    public static class PREFIX {
-        /** 시스템 로그 */
-        public final static String LOG = "[ DMGR-Core ] ";
-        /** 일반 */
-        public final static String CHAT = "§3§l[ §b§lDMGR §3§l] §f";
-        /** 경고 */
-        public final static String CHAT_WARN = "§3§l[ §b§lDMGR §3§l] §c";
     }
 }
